Moved abstract jct compiler out of the impl package
